moved some files to src directory. added abstractions.ts file.
