+++ src/abstractions.ts
@@ -1,0 +1,3 @@
+export interface Navigation {
+    [sectionName: string]: Array<string | object>;
+}
